+++ clients/ctl/admin-ui/src/features/data-subjects/data-subject.slice.ts
@@ -45,6 +45,14 @@ export const dataSubjectsApi = createApi({
       }),
       invalidatesTags: ["Data Subjects"],
     }),
+    deleteDataSubject: build.mutation<string, string>({
+      query: (key) => ({
+        url: `data_subject/${key}`,
+        params: { resource_type: "data_subject" },
+        method: "DELETE",
+      }),
+      invalidatesTags: ["Data Subjects"],
+    }),
   }),
 });
 
@@ -52,6 +60,7 @@ export const {
   useGetAllDataSubjectsQuery,
   useUpdateDataSubjectMutation,
   useCreateDataSubjectMutation,
+  useDeleteDataSubjectMutation,
 } = dataSubjectsApi;
 
 export const dataSubjectsSlice = createSlice({