+++ clients/ctl/admin-ui/src/features/data-use/data-use.slice.ts
@@ -45,6 +45,14 @@ export const dataUseApi = createApi({
       }),
       invalidatesTags: ["Data Uses"],
     }),
+    deleteDataUse: build.mutation<string, string>({
+      query: (key) => ({
+        url: `data_use/${key}`,
+        params: { resource_type: "data_use" },
+        method: "DELETE",
+      }),
+      invalidatesTags: ["Data Uses"],
+    }),
   }),
 });
 
@@ -52,6 +60,7 @@ export const {
   useGetAllDataUsesQuery,
   useUpdateDataUseMutation,
   useCreateDataUseMutation,
+  useDeleteDataUseMutation,
 } = dataUseApi;
 
 export const dataUseSlice = createSlice({