+++ clients/ctl/admin-ui/src/features/taxonomy/hooks.tsx
@@ -19,33 +19,38 @@ import {
 } from "../common/form/inputs";
 import {
   useCreateDataQualifierMutation,
+  useDeleteDataQualifierMutation,
   useGetAllDataQualifiersQuery,
   useUpdateDataQualifierMutation,
 } from "../data-qualifier/data-qualifier.slice";
 import {
   useCreateDataSubjectMutation,
+  useDeleteDataSubjectMutation,
   useGetAllDataSubjectsQuery,
   useUpdateDataSubjectMutation,
 } from "../data-subjects/data-subject.slice";
 import {
   useCreateDataUseMutation,
+  useDeleteDataUseMutation,
   useGetAllDataUsesQuery,
   useUpdateDataUseMutation,
 } from "../data-use/data-use.slice";
 import {
   useCreateDataCategoryMutation,
+  useDeleteDataCategoryMutation,
   useGetAllDataCategoriesQuery,
   useUpdateDataCategoryMutation,
 } from "./taxonomy.slice";
 import type { FormValues } from "./TaxonomyFormBase";
-import { Labels, TaxonomyEntity, TaxonomyRTKResult } from "./types";
+import { Labels, RTKResult, TaxonomyEntity } from "./types";
 
 export interface TaxonomyHookData<T extends TaxonomyEntity> {
   data?: TaxonomyEntity[];
   isLoading: boolean;
   labels: Labels;
-  edit: (entity: T) => TaxonomyRTKResult;
-  onCreate: (entity: T) => TaxonomyRTKResult;
+  handleCreate: (entity: T) => RTKResult<TaxonomyEntity>;
+  handleEdit: (entity: T) => RTKResult<TaxonomyEntity>;
+  handleDelete: (key: string) => RTKResult<string>;
   extraFormFields?: ReactNode;
   transformEntityToInitialValues: (entity: T) => FormValues;
 }
@@ -74,15 +79,17 @@ export const useDataCategory = (): TaxonomyHookData<DataCategory> => {
     parent_key: "Parent category",
   };
 
-  const [edit] = useUpdateDataCategoryMutation();
-  const [onCreate] = useCreateDataCategoryMutation();
+  const [handleEdit] = useUpdateDataCategoryMutation();
+  const [handleDelete] = useDeleteDataCategoryMutation();
+  const [handleCreate] = useCreateDataCategoryMutation();
 
   return {
     data,
     isLoading,
     labels,
-    edit,
-    onCreate,
+    handleCreate,
+    handleEdit,
+    handleDelete,
     transformEntityToInitialValues: transformTaxonomyBaseToInitialValues,
   };
 };
@@ -104,7 +111,7 @@ export const useDataUse = (): TaxonomyHookData<DataUse> => {
   };
 
   const [edit] = useUpdateDataUseMutation();
-  const [onCreate] = useCreateDataUseMutation();
+  const [create] = useCreateDataUseMutation();
 
   const transformFormValuesToEntity = (formValues: DataUse) => ({
     ...formValues,
@@ -116,11 +123,12 @@ export const useDataUse = (): TaxonomyHookData<DataUse> => {
         : formValues.legitimate_interest_impact_assessment,
   });
 
+  const [handleDelete] = useDeleteDataUseMutation();
   const handleEdit = (entity: DataUse) =>
     edit(transformFormValuesToEntity(entity));
 
   const handleCreate = (entity: DataUse) =>
-    onCreate(transformFormValuesToEntity(entity));
+    create(transformFormValuesToEntity(entity));
 
   const transformEntityToInitialValues = (du: DataUse) => {
     const base = transformTaxonomyBaseToInitialValues(du);
@@ -168,8 +176,9 @@ export const useDataUse = (): TaxonomyHookData<DataUse> => {
     data,
     isLoading,
     labels,
-    edit: handleEdit,
-    onCreate: handleCreate,
+    handleCreate,
+    handleEdit,
+    handleDelete,
     extraFormFields,
     transformEntityToInitialValues,
   };
@@ -188,7 +197,8 @@ export const useDataSubject = (): TaxonomyHookData<DataSubject> => {
   };
 
   const [edit] = useUpdateDataSubjectMutation();
-  const [onCreate] = useCreateDataSubjectMutation();
+  const [create] = useCreateDataSubjectMutation();
+  const [handleDelete] = useDeleteDataSubjectMutation();
 
   const transformFormValuesToEntity = (entity: TaxonomyEntity) => {
     const transformedEntity = {
@@ -209,7 +219,7 @@ export const useDataSubject = (): TaxonomyHookData<DataSubject> => {
     edit(transformFormValuesToEntity(entity));
 
   const handleCreate = (entity: TaxonomyEntity) =>
-    onCreate(transformFormValuesToEntity(entity));
+    create(transformFormValuesToEntity(entity));
 
   const transformEntityToInitialValues = (ds: DataSubject) => {
     const base = transformTaxonomyBaseToInitialValues(ds);
@@ -241,8 +251,9 @@ export const useDataSubject = (): TaxonomyHookData<DataSubject> => {
     data,
     isLoading,
     labels,
-    edit: handleEdit,
-    onCreate: handleCreate,
+    handleCreate,
+    handleEdit,
+    handleDelete,
     extraFormFields,
     transformEntityToInitialValues,
   };
@@ -258,15 +269,17 @@ export const useDataQualifier = (): TaxonomyHookData<DataQualifier> => {
     parent_key: "Parent data qualifier",
   };
 
-  const [edit] = useUpdateDataQualifierMutation();
-  const [onCreate] = useCreateDataQualifierMutation();
+  const [handleCreate] = useCreateDataQualifierMutation();
+  const [handleEdit] = useUpdateDataQualifierMutation();
+  const [handleDelete] = useDeleteDataQualifierMutation();
 
   return {
     data,
     isLoading,
     labels,
-    edit,
-    onCreate,
+    handleCreate,
+    handleEdit,
+    handleDelete,
     transformEntityToInitialValues: transformTaxonomyBaseToInitialValues,
   };
 };